refactor: search dialog to use native dialog, add polyfill for safari closeby suppor
