+++ src/utils/booleans.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line unicorn/prefer-global-this
+export const isBrowser = typeof window !== "undefined";
